requirements.txt

# Additional requirements go here
django-jenkins==0.110.0
djangorestframework==3.6.3
